The first step towards an Ironic Teeth driver
TeethPower still needs to be implemented. I'm combining the prepare_image and
run_image commands to the agent. The design here is that the agent will boot
via PXE, and check in with the driver to see what it needs to do. We control
this with node.provision_state and node.target_provision_state. The node will
be told it needs to decom, cache, or run an image. The reason for the change
is conductor doesn't seem to be running on gevent, so async is out. We still
need to write an API controller to let the node talk to the driver.